fix: only check for per-thread rusage in the benchmark that requires it
The `single_row_throughput_benchmark` doesn't use `getrusage()`, so
checking for it in `Config` unnecessarily precludes it from running
(i.e. when using `bazel` builds).

Instead, check in `multiple_rows_cpu_benchmark` which actually does care.
